requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
pandas>=1.0.0
beautifulsoup4>=4.9.0